refactor: split region map unit markers into dedicated part (#3878)
