+++ app/lib/features/game/flame/region_map/region_map_component.dart
@@ -33,6 +33,7 @@ part 'region_map_component_render_political_labels.dart';
 part 'region_map_component_render_political_borders.dart';
 part 'region_map_component_render_markers.dart';
 part 'region_map_component_render_markers_settlements.dart';
+part 'region_map_component_render_markers_units.dart';
 
 final _log = packageLogger();
